requirements.txt
# Core Dependencies
python>=3.8

# Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# MLflow and Databricks
mlflow>=2.8.0
databricks-cli>=0.18.0
databricks-sql-connector>=2.9.0

# Additional ML Libraries
xgboost>=1.7.0
lightgbm>=3.3.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
joblib>=1.1.0
click>=8.0.0

# Logging and Monitoring
loguru>=0.6.0

# API and Deployment
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

